backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
openai==1.30.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
